move Namespace to utils
